{"schema":"decodex/commit/1","summary":"Split MCP autonomy server modules","authority":"XY-1115"}
