+++ apps/decodex/src/mcp/planning/autonomy.rs
@@ -1,3 +1,5 @@
 mod args;
+mod objective;
+mod proposal;
 mod results;
-mod server;
+mod signal;
